Explicitly apply seek timestamp on native
The mini player doesn't seem to need it, but when seeking using the
touch bar (I can't seem to get it to show up again for some reason...)
it didn't seem to use the seek it just applied—hopefully this fixes
that.

+++ macos/gakki/gakki/Model/Command.swift
@@ -30,7 +30,7 @@ struct Command: Codable {
 
     // .setState:
     var state: State?
-    var currentTime: Float?
+    var currentTime: Double?
 
     // .setNowPlaying:
     var title: String?
